feat : 클럽 멤버 인증용 서비스 분리
